Support deferring the start of logux client.
It support by calling store.startLoguxClient or dispatching a logux/start action with config.
e.g. dispatch{{ type: 'logux/start', config: {} }}

A logux/started action will be dispatched to indicate the client is started.
If logux-reducer is used, the state will have the flag started = true.
If using store.startLoguxClient, the return value is a promise resolve when the client is started.

The enhancer support a callback of onClientStart, which can be used to config logux-status.

It sounds better to put when creating the enhancer since the code triggering logux/start is likely more in application logic.